+++ src/app.rs
@@ -2,15 +2,30 @@ use leptos::*;
 use leptos_meta::*;
 use leptos_router::*;
 
-use crate::frontend::{
-    components::Sidebar,
-    views::{Chat, Home},
+use crate::{
+    frontend::{
+        components::Sidebar,
+        views::{Chat, Home},
+    },
+    models,
 };
+
+#[derive(Copy, Clone)]
+pub struct MessagesContext {
+    pub messages: ReadSignal<Vec<models::Message>>,
+    pub set_messages: WriteSignal<Vec<models::Message>>,
+}
 
 #[component]
 pub fn App() -> impl IntoView {
     // Provides context that manages stylesheets, titles, meta tags, etc.
     provide_meta_context();
+
+    let (messages, set_messages) = create_signal(Vec::<models::Message>::new());
+    provide_context(MessagesContext {
+        messages,
+        set_messages,
+    });
 
     view! {
         <Title text="Welcome to LokAI!"/>
