+++ src/frontend/components/prompt.rs
@@ -1,0 +1,6 @@
+use leptos::*;
+
+#[component]
+pub(crate) fn Prompt() -> impl IntoView {
+    todo!()
+}
